backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.36.0
torch==2.1.0
deep-translator==1.11.4
langdetect==1.0.9
easyocr==1.7.1
Pillow==10.1.0
numpy==1.26.2
scipy==1.11.4
